clean up v1 tests
